cyber_security_smart_lighting/backend/requirements.txt
# backend/requirements.txt

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0

# LangChain and LangGraph for AI Agents
langchain==0.1.0
langchain-groq==0.0.1
langgraph==0.0.26
langchain-community==0.0.10

# Groq API client
groq==0.4.1

# For async operations
asyncio==3.4.3
aiofiles==23.2.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Date and time handling
python-dateutil==2.8.2

# For development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional: for enhanced logging
loguru==0.7.2

# Optional: for monitoring
prometheus-client==0.19.0

# CORS support (included with FastAPI but explicit for clarity)
fastapi-cors==0.0.6